After @eroell review
