+++ tests/anndata/test_anndata_ext.py
@@ -20,6 +20,8 @@ from ehrapy.anndata.anndata_ext import (
     move_to_obs,
     move_to_x,
     set_numeric_vars,
+    _are_ndarrays_equal,
+    _is_val_missing
 )
 from tests.conftest import TEST_DATA_PATH
 
@@ -500,3 +502,17 @@ def test_set_numeric_vars(adata_strings_encoded):
 
     with pytest.raises(NotEncodedError, match=r"not yet been encoded"):
         set_numeric_vars(adata_strings, values)
+
+
+def test_are_ndarrays_equal(impute_num_adata):
+    impute_num_adata_copy = impute_num_adata.copy()
+    assert _are_ndarrays_equal(impute_num_adata.X, impute_num_adata_copy.X)
+    impute_num_adata_copy.X[0, 0] = 42.0
+    assert not _are_ndarrays_equal(impute_num_adata.X, impute_num_adata_copy.X)
+
+
+def test_is_val_missing(impute_num_adata):
+    assert np.array_equal(
+        _is_val_missing(impute_num_adata.X),
+        np.array([[False, False, True], [False, False, False], [True, False, False], [False, False, True]]),
+    )
